requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pandas==2.2.1
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.120.0
gspread==6.0.0
gspread-dataframe==3.3.1
flask==2.3.3
flask-cors==4.0.0 